remove dead code, refactor RK functions to separate file
