+++ src/filter/mod.rs
@@ -7,6 +7,7 @@ pub mod linear;
 pub mod angles;
 pub mod jacobian;
 pub mod utils;
+pub mod runge_kutta;
 
 pub mod filter_gain;
 pub mod filter_means;
