+++ src/macros.rs
@@ -27,24 +27,3 @@ macro_rules! def_constant {
         )+
     };
 }
-
-#[macro_export]
-/// Quickly generate test data
-macro_rules! test_data {
-    ($len:expr; $($name:ident, $type:ident),+) => {
-        $(
-            let mut $name = Vec::with_capacity($len);
-            let var : $type = $type::new_random();
-            for _ in 0..$len {
-                $name.push(var.clone());
-            }
-        )+
-    };
-    (mat; $len:expr; $name:ident, $type:ident) => {
-        let mut $name = Vec::with_capacity($len);
-        for _ in 0..$len {
-            let default = $type::default();
-            $name.push(default)
-        }
-    };
-}
